Refactor Map to separate folder
